Conflicts:
	geekway-model/src/main/java/com/bruce/geekway/model/wx/pay/WxDeliveryTemplate.java
	geekway-pay/src/main/java/com/bruce/geekway/controller/wx/WxProductOrderController.java
	geekway-pay/src/main/resources/deliverTemplate.xml
	geekway-pay/src/main/webapp/WEB-INF/jsp/order/buyNow.jsp
	geekway-service/src/main/java/com/bruce/geekway/service/product/IWxDeliveryTemplateService.java
	geekway-service/src/main/java/com/bruce/geekway/service/product/impl/WxDeliveryTemplateServiceImpl.java
	geekway-service/src/main/java/com/bruce/geekway/service/product/impl/WxProductSkuServiceImpl.java

+++ geekway-model/src/main/java/com/bruce/geekway/model/wx/pay/WxDeliveryTemplate.java
@@ -134,7 +134,7 @@ public class WxDeliveryTemplate {
 			this.addFees = addFees;
 		}
 
-		public int getStartStandards() {
+		public double getStartStandards() {
 			return startStandards;
 		}
 
@@ -146,7 +146,7 @@ public class WxDeliveryTemplate {
 			return startFees;
 		}
 
-		public void setStartFees(double startFees) {
+		public void setStartFees(int startFees) {
 			this.startFees = startFees;
 		}
